+++ src/main/java/com/example/ajouevent_be_v2/scheduler/TokenValidationScheduler.java
@@ -1,4 +1,4 @@
-package com.example.ajouevent_be_v2.service.webhook;
+package com.example.ajouevent_be_v2.scheduler;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -11,6 +11,7 @@ import java.util.stream.Collectors;
 
 import com.example.ajouevent_be_v2.domain.member.Token;
 import com.example.ajouevent_be_v2.service.token.TokenService;
+import com.example.ajouevent_be_v2.service.webhook.FcmPushService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
